+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Add support for the ``allow_custom_choice`` field property in Python
+  processes
+
+
 ===================
 21.0.0 - 2020-03-16
 ===================